requirements.txt
python-telegram-bot[ext]
SQLAlchemy
requests
httpx
python-dotenv